requirements.txt
streamlit
pandas
numpy
matplotlib
seaborn
plotly.express
scikit-learn
joblib
altair